Make 'allocate_hexagon_array' a hexagon contrib API
