+++ tests/python/contrib/test_hexagon/topi/slice_op/test_cast_slice.py
@@ -22,8 +22,9 @@ import tvm
 import tvm.testing
 from tvm import te
 import tvm.topi.hexagon.slice_ops as sl
+from tvm.contrib.hexagon import allocate_hexagon_array
 
-from ...infrastructure import allocate_hexagon_array, transform_numpy, get_hexagon_target
+from ...infrastructure import transform_numpy, get_hexagon_target
 
 
 class TestCastF16F32Slice2d:
