+++ tests/python/contrib/test_hexagon/topi/slice_op/test_clip_slice.py
@@ -22,8 +22,9 @@ import numpy as np
 from tvm import te
 import tvm.testing
 import tvm.topi.hexagon.slice_ops as sl
+from tvm.contrib.hexagon import allocate_hexagon_array
 
-from ...infrastructure import allocate_hexagon_array, transform_numpy, get_hexagon_target
+from ...infrastructure import transform_numpy, get_hexagon_target
 
 input_layout = tvm.testing.parameter(
     "nhwc-8h2w32c2w-2d",
