+++ tests/python/contrib/test_hexagon/topi/test_add_subtract_multiply.py
@@ -21,8 +21,8 @@ import tvm
 from tvm import te
 import tvm.topi.hexagon.slice_ops as sl
 import tvm.topi.hexagon.qnn as qn
+from tvm.contrib.hexagon import allocate_hexagon_array
 from ..infrastructure import (
-    allocate_hexagon_array,
     transform_numpy,
     quantize_np,
     get_hexagon_target,
